Put backend implementation for thread load/store in the correct place
